Moved make_molecule_from_file to prevent cyclic import
