+++ functions.py
@@ -18,13 +18,14 @@ verified_input: Verify that the user has input a value which can be converted to
 
 """
 
+from os import getcwd, makedirs, path
+
 import matplotlib
 import matplotlib.pyplot as plt
 import numpy as np
 import pyquaternion as pq
 
 from classes import Atom, Molecule
-from parsing import make_output_folder
 
 
 def generate_figure(mo: Molecule):
@@ -255,6 +256,10 @@ def check_bonds(m1, m2):
     Only a very complex scenario would really defeat this detection method.
     A complex problem == complex solution, thus, I procrastinate.
     """
+    # Update bond graphs
+    m1.make_bond_graph()
+    m2.make_bond_graph()
+
     # Makes a list of the bond counts for each atom
     m1struct = [len(m1.bonds[b]) for b in m1.bonds]
     m2struct = [len(m2.bonds[b]) for b in m2.bonds]
@@ -267,7 +272,7 @@ def check_bonds(m1, m2):
 
 
 def randomly_orient(mo):
-    """Randomly orient a given molecule in 3d space"""
+    """Randomly orient a given molecule in 3d space using uniform distributions"""
     mo = center_on_atom(mo, 0)
     new_mo = Molecule(name=mo.name, atoms=list())
 
@@ -282,3 +287,32 @@ def randomly_orient(mo):
         new_mo.add_atom(new_atom)
 
     return new_mo
+
+
+def make_output_folder(sub: str = "") -> str:
+    """
+    Makes a directory in the script location to output the downloaded files
+
+    Parameters
+    ----------
+    sub: The name of the directory to be made.
+
+    Returns
+    -------
+    dir_path: The directory pointing to :sub:
+
+    """
+    # Finds the current directory
+    dir_path = getcwd()
+
+    # Makes the path for the new folder
+    dir_path = dir_path + fr"\{sub}"
+
+    # If the folder doesn't exist, make it.
+    if not path.exists(dir_path):
+        try:
+            makedirs(dir_path)
+        except FileExistsError:
+            # Sometimes this error pops when using threading or multiprocessing.
+            pass
+    return dir_path
